requirements.txt
pycryptodome==3.12.0
rich==11.0.0
